+++ sort-algorithms/search-headers/CC-BINARY-02.h
@@ -1,17 +1,17 @@
-int _binarySearch(int* Array, int query)
+long int _binarySearch(long int *Array, long int query)
 {
-     int begin = 0; 
-     int end = length-1;
-     int middle;
+     long int begin = 0;
+     long int end = length - 1;
+     long int middle;
      while (begin <= end)
      {
-          middle = (begin + end)/2;
+          middle = (begin + end) / 2;
           if (query == Array[middle])
                return middle;
           else if (query < Array[middle])
-               end = middle-1;
+               end = middle - 1;
           else
-               begin = middle+1;
+               begin = middle + 1;
      }
-     return -1;   // not found
+     return -1; // not found
 }
